address /simplify: reuse getProcessesList, guard log-fetch race, de-dup
- Reuse the existing apiCommands.getProcessesList for the PM2 picker instead
  of a duplicate getLogProcesses wrapper (apiLogs keeps only getProcessLogs).
- Add a monotonic request-id guard so a superseded/post-close log fetch can't
  clobber newer logs; fetch the process list once per open.
- Relocate the run-domain runLogProcessName helper out of the generic
  ProcessLogModal into RunsHistoryPage; derive the process name once.
- Drop the redundant open-guard and stale effect dep.
