+++ client/src/components/ui/ProcessLogModal.test.jsx
@@ -2,28 +2,17 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent, waitFor, within } from '@testing-library/react';
 
 vi.mock('../../services/api', () => ({
-  getLogProcesses: vi.fn(),
+  getProcessesList: vi.fn(),
   getProcessLogs: vi.fn(),
 }));
 
-import { getLogProcesses, getProcessLogs } from '../../services/api';
-import ProcessLogModal, { runLogProcessName } from './ProcessLogModal';
-
-describe('runLogProcessName', () => {
-  it('maps cos-agent runs to the portos-cos process', () => {
-    expect(runLogProcessName('cos-agent')).toBe('portos-cos');
-  });
-
-  it('maps devtools and unknown sources to the main portos-server process', () => {
-    expect(runLogProcessName('devtools')).toBe('portos-server');
-    expect(runLogProcessName(undefined)).toBe('portos-server');
-  });
-});
+import { getProcessesList, getProcessLogs } from '../../services/api';
+import ProcessLogModal from './ProcessLogModal';
 
 describe('ProcessLogModal', () => {
   beforeEach(() => {
     vi.clearAllMocks();
-    getLogProcesses.mockResolvedValue([{ name: 'portos-server' }, { name: 'portos-cos' }]);
+    getProcessesList.mockResolvedValue([{ name: 'portos-server' }, { name: 'portos-cos' }]);
     getProcessLogs.mockResolvedValue({ processName: 'portos-server', lines: 200, logs: 'boot ok\nerror: boom' });
   });
 
@@ -53,7 +42,7 @@ describe('ProcessLogModal', () => {
   });
 
   it('keeps the hinted process selectable even when PM2 does not report it', async () => {
-    getLogProcesses.mockResolvedValueOnce([{ name: 'portos-cos' }]);
+    getProcessesList.mockResolvedValueOnce([{ name: 'portos-cos' }]);
     render(<ProcessLogModal open onClose={() => {}} processName="portos-server" />);
     await waitFor(() => expect(getProcessLogs).toHaveBeenCalledWith('portos-server', 200));
     const select = screen.getByLabelText('Process');
